diverse Aktualisierungen und Anpassungen (index.php, header.php, usw), Funktionen bei newsbeitraege_erstellen.php, usw

+++ Übungen_MF.php
@@ -17,7 +17,7 @@ if (isset($_SESSION['role']) && $_SESSION['role'] !== "admin") {
 
   <title>Newsbeiträge erstellen</title>
 
-  <link href="./css_Daten/colors.css" rel="stylesheet">
+  <link href="./css_Daten/colors_MF.css" rel="stylesheet">
 
 </head>
 
@@ -37,18 +37,35 @@ if (isset($_SESSION['role']) && $_SESSION['role'] !== "admin") {
     $_SERVER["REQUEST_METHOD"] === "POST"
     && isset($_FILES["file"])
   ) {
-    if (move_uploaded_file(
-      $_FILES["file"]["tmp_name"],
-      $uploadDir . $_FILES["file"]["name"]
-    )) {
-      //      echo "<p class='green'>";
-      echo "Datei wurde erfolgreich\n hochgeladen!<br />";
-      //      echo "TEST";
+    //folgender Teil schränkt auf .pdf ein
+    if ($_FILES['file']['type'] == "application/pdf") {
+      if ($_FILES['file']['size'] > 500000) {
+        echo "<p class='red'>";
+        echo "ACHTUNG - Datei zu groß - max. 500kB sind erlaubt!";
+      } else {
+        if (file_exists($uploadDir . $_FILES["file"]["name"])) {
+          echo "<p class='red'>";
+          echo "ACHTUNG - diese Datei ist schon vorhanden!";
+        } else {
+          if (move_uploaded_file(
+            $_FILES["file"]["tmp_name"],
+            $uploadDir . $_FILES["file"]["name"]
+          )) {
+            echo "<p class='green'>";
+            echo "Die Datei ";
+            echo $_FILES["file"]["name"];
+            echo " wurde erfolgreich hochgeladen!<br />";
 
-      echo "<br>";
+            echo "<br>";
+          } else {
+            echo "Fehler beim Upload!<br />";
+            //      echo "<p class='red' Fehler beim Upload!\n"; -> prüfen ob 'red' funktionieren würde!!!
+          }
+        }
+      }
     } else {
-      echo "Fehler beim Upload!<br />";
-      //      echo "<p class='red' Fehler beim Upload!\n"; -> prüfen ob 'red' funktionieren würde!!!
+      echo "<p class='red'>";
+      echo "Leider kein pdf, bitte versuchen Sie es nochmals!<br />";
     }
   }
 
@@ -71,7 +88,12 @@ if (isset($_SESSION['role']) && $_SESSION['role'] !== "admin") {
             <label for="file" class="form-label">
               <h4>Bitte hier Daten hochladen</h4>
             </label>
-            <input class="form-control" type="file" id="file" name="file">
+            <!-- Angabe einer maximalen Dateigröße in Bytes (hier: 500 kB). 
+            <input type="hidden" name="MAX_FILE_SIZE" value="500000"> 
+            Macht KEINEN Sinn, da keine Fehlermeldung kommt, daher in php schreiben -->
+            <label>Wählen sie bitte ein pdf-Dokument aus zum Hochladen aus!
+              <input class="form-control" type="file" id="file" name="file" accept=".pdf">
+            </label>
             <button class="w-100 btn btn-lg btn-primary btn-brown" type="submit">Upload</button>
           </div>
         </form>
@@ -92,9 +114,9 @@ if (isset($_SESSION['role']) && $_SESSION['role'] !== "admin") {
                 //               echo "<li>" . $files[$i] . "</li>";
               }
               if (count($files) == 2) {
-                echo "Keine Files vorhanden ...";
-//                echo "<li>Keine Files vorhanden ...</li>";
-//                echo "<p class='red' <li>Keine Files vorhanden ...</li>";
+                //                echo "Keine Files vorhanden ...";
+                //                echo "<li>Keine Files vorhanden ...</li>";
+                echo "<p class='red' <li>Keine Files vorhanden ...</li>";
               }
             }
